DrawJack Game
In this Basic Python Project, I combine the Blackjack game with simple drawing card game. In the beginning you and the cpu automatically draw two cards and then you start comparing the sum of the numbers on your cards with the cpu. You keep drawing and nearest to the number 21 wins and if you exceed 21, you lose. You can draw a card or stay with yours during your turn. Cpu makes his decision on his own. This project is very suitable for development and you can share your developments with me.